Revert "コンポーネントとロジックの分離STEP1"
