+++ web/app/room/[roomId]/page.tsx
@@ -1,4 +1,4 @@
-import Room from "@/features/room/Room";
+import Room from "@/components/Room";
 import type { GameRoom } from "@/types/room";
 import { cookies, headers } from "next/headers";
