+++ web/components/modals/GameResultModal.tsx
@@ -66,7 +66,7 @@ export default function GameResultModal({
       ref={ref}
     >
       <div
-        className={`${animation} grid place-items-center gap-4 mx-1 backdrop:bg-black/80 p-6 text-card-foreground shadow-sm bg-gray-800 border-2 ${borderColor}`}
+        className={`${animation} grid place-items-center gap-4 backdrop:bg-black/80 p-6 text-card-foreground shadow-sm w-full bg-gray-800 border-2 ${borderColor}`}
       >
         <div className="flex items-center flex-col gap-4">
           <h2 className="font-semibold text-red-500">ゲーム終了</h2>
