--- web/utils/copyToClipboard.ts
@@ -1,8 +1,0 @@
-export const copyToClipboard = async (text: string) => {
-  try {
-    await navigator.clipboard.writeText(text);
-  } catch (error) {
-    console.error(error);
-    throw new Error("Failed to copy ID");
-  }
-};
