+++ web/utils/toast/Toast.tsx
@@ -11,7 +11,7 @@ export function Toast() {
   const { message, isOpen } = context;
 
   return isOpen ? (
-    <div className="flex justify-center p-4 bg-gray-800 border-2 border-red-500 text-white rounded-lg shadow-lg">
+    <div className="fixed p-4 mx-1 bg-gray-800 border-2 text-white rounded-lg shadow-lg">
       {message}
     </div>
   ) : null;
